refactor(ui): move PatchItem to the only file where it is used
